add the newlib-nano stdlib

+++ 022task_scheduler_from_scratch/stm32_startup.c
@@ -221,7 +221,7 @@ uint32_t vectors[] __attribute__ ((section(".isr_vector"))) = {
 
 void Reset_Handler(void) {
 	// copy .data section to SRAM
-	uint32_t size = &_edata - &_sdata;
+	uint32_t size = (uint32_t)&_edata - (uint32_t)&_sdata;
 	uint8_t *pDst = (uint8_t*)&_sdata; // sram
 	uint8_t *pSrc = (uint8_t*)&_etext; // flash
